Fix article list item prop types
