+++ packages/article-list/src/article-list-item-prop-types.js
@@ -1,10 +1,39 @@
 import PropTypes from "prop-types";
-import * as baseTypes from "./article-list-item-prop-types.base";
 
-export const propTypes = {
-  ...baseTypes.propTypes,
-  onPress: PropTypes.func.isRequired
+export const propTypesBase = {
+  article: PropTypes.shape({
+    headline: PropTypes.string,
+    id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
+    label: PropTypes.string,
+    leadAsset: PropTypes.shape({
+      crop: PropTypes.shape({
+        url: PropTypes.string
+      }),
+      title: PropTypes.string
+    }),
+    longSummary: PropTypes.array,
+    publicationName: PropTypes.string,
+    publishedTime: PropTypes.string,
+    shortHeadline: PropTypes.string,
+    shortSummary: PropTypes.array,
+    summary: PropTypes.array,
+    url: PropTypes.string
+  }),
+  fadeImageIn: PropTypes.bool,
+  highResSize: PropTypes.number
 };
 
-// eslint-disable-next-line prefer-destructuring
-export const defaultProps = baseTypes.defaultProps;
+export const propTypes = {
+  ...propTypesBase,
+  imageRatio: PropTypes.number,
+  showImage: PropTypes.bool
+};
+
+export const defaultProps = {
+  article: null,
+  fadeImageIn: false,
+  highResSize: null,
+  imageRatio: 1,
+  showImage: true,
+  summary: []
+};
