--- packages/article-list/src/article-list-item-prop-types.web.js
@@ -1,1 +1,0 @@
-export { propTypes, defaultProps } from "./article-list-item-prop-types.base";
